introduce generic ValueSignal type
