+++ packages/java/endpoint/src/main/java/com/vaadin/hilla/signals/core/registry/SignalsRegistry.java
@@ -1,7 +1,6 @@
 package com.vaadin.hilla.signals.core.registry;
 
-import com.vaadin.hilla.signals.NumberSignal;
-
+import com.vaadin.hilla.signals.ValueSignal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -20,7 +19,7 @@ public final class SignalsRegistry {
 
     private static final Logger LOGGER = LoggerFactory
             .getLogger(SignalsRegistry.class);
-    private final Map<UUID, NumberSignal> signals = new WeakHashMap<>();
+    private final Map<UUID, ValueSignal<?>> signals = new WeakHashMap<>();
     private final Map<String, UUID> clientSignalToSignalMapping = new HashMap<>();
 
     SignalsRegistry() {
@@ -42,7 +41,7 @@ public final class SignalsRegistry {
      *             if {@code clientSignalId} or {@code signal} is null
      */
     public synchronized void register(String clientSignalId,
-            NumberSignal signal) {
+                                      ValueSignal<?> signal) {
         Objects.requireNonNull(clientSignalId,
                 "Client signal id must not be null");
         Objects.requireNonNull(signal, "Signal must not be null");
@@ -68,7 +67,7 @@ public final class SignalsRegistry {
      * @throws NullPointerException
      *             if {@code clientSignalId} is null
      */
-    public synchronized NumberSignal get(String clientSignalId) {
+    public synchronized ValueSignal<?> get(String clientSignalId) {
         Objects.requireNonNull(clientSignalId,
                 "Client signal id must not be null");
         UUID signalId = clientSignalToSignalMapping.get(clientSignalId);
@@ -92,7 +91,7 @@ public final class SignalsRegistry {
      * @throws NullPointerException
      *             if {@code signalId} is null
      */
-    public synchronized NumberSignal getBySignalId(UUID signalId) {
+    public synchronized ValueSignal<?> getBySignalId(UUID signalId) {
         Objects.requireNonNull(signalId, "Signal id must not be null");
         return signals.get(signalId);
     }
